Bonus: Move the QaCollapsiblePanel, which was only used in one component
