+++ src/components/qaSection/QaCollapsiblePanel.tsx
